+++ Test Bench/Header Files/Components/NoRandomComponent.h
