+++ Test Bench/main.cpp
@@ -4,7 +4,6 @@
 
 #include <thread>
 
-#include "Header Files/MovableObject.h"
 #include "Header Files/MovableAnimatedObject.h"
 #include "Header Files/AnimationChangerSystem.h"
 #include "Header Files/Level.h"
